bot/requirements.txt
aiogram==2.25.2
redis
python-dotenv
